Split Exchange trait in two
The Exchange trait had two functions: hiding how the exchange protocol 
works (just the createHandshake method) and holding descriptive 
information about an exchange (the rest of the methods, getters in fact).

The former function is kept in an Exchange case class and the former
into a new trait named ExchangeProtocol.