Several improvements on UI order submission form
* The price model was inverted. The limit order should be enabled and the
market-price order should be disabled. It was the other way around, what has
been fixed.
* Use a custom `DecimalNumberTextField` that only accepts a decimal number
as input. This custom control implements the necessary event filters and
property listeners to forbid inputs that doesn't match a double number. This
new control is used for all text fields where a decimal number is expected.
* Inject a function to process the form submit. By default, this only shows
a dialog informing the order was submitted.
* Remove innecesary fields for custom priced orders.
* Require a valid bitcoin amount to enable the _submit_ button.
* Adecuate some labels to finantial jargon
* Some style issues fixed.