Consider Order class as a domain model type
The order class was placed in `protocol.messages.brokerage` package, since it
was considered as part of a protocol message. But this type is actually
representing an entity of the Coinffeine domain model. Thus, the class has been
moved to the `common` package until we have a better place to store our
domain model classes.