Redesign of operations UI
The UI mock design considers that order submission should be a
sub-view of the operations view. Considering that this view
is only used to introduce some inputs and submit a new request,
it makes more sense to design it as a form that opens in a new
modal window. This form would be accessible from a button in a
corner of the operations view.