Make `Exchange.Id` an outer class `ExchangeId`
The current classes in `model` module representing the exchange are not model
classes but a representation of the exchange state from the protocol
perspective. As a previos step before moving such classes to the `peer` package,
the exchange identifier, which is actually an entity of the model, should be
extracted from the protocol-view exchange representation.