Remove dependency from `Both` to `Role`
The `Role` class is indeed a piece of information of how the protocol works, not
the model. The `Both` class, that is used almost everywhere as a utility to pair
buyer and seller data, depends on `Role`. This change removes such a dependency
by moving the role-related functions to `Role`.