Move back `Exchange` class hierarchy into `coinffeine.model`
After working with the idea to have a from-scratch representation of exchanges
in `coinffeine.model`, we reach a code with just minor differences respect
the classes that were moved to `coinffeine.peer.protocol`. This change moves
the classes back to its original emplacement to be used as domain model
representation of an exchange.