Consider order cancellation as a new order status
So far, an order cancellation meant the order was removed from the system. A
`OrderCancelledEvent` was sent, and the order disappeared from the orders list
in the operations view.

This change implements a different approach that consist in marking the order
with a new status of `CancelledOrder`. Instead of sending a
`OrderCancelledEvent`, the new state of the `Order` object is calculated and
propagated with a `OrderUpdatedEvent`. The UI now is adapted to this new
behavior, enabling the "cancel" button only in case the selected order is
cancellable.