Fix build issues while connected to OpenVPN
While connected with OpenVPN, TomP2P inters that the public IP address
corresponds to that connects to the VPN. That prevents the tests to pass since
both peers are in the same subnet and cannot communicate among them using
the VPN gateway.

This change fixes that problem by making `ProtoMessageGateway` to accept a
sequence of network interfaces to ignore in its initialization. For testing,
this list includes `utun0`, the name OpenVPN receives in OSX. For production,
the list is obtained from the setting `coinffeine.peer.ifaces.ignore` that
comes from the config file.