Fix UDP socket limit issue while passing tests
In some boxes, the maximum number of UDP sockets is limited to a pretty low
value. Too low for TomP2P to work. The property `sun.net.maxDatagramSockets`
is set to a value of 128 in SBT to avoid such issue. This requires the tests
to be executed in fork mode, which in turn requires the IP addresses of
`DefaultTcpPortAllocator` to be assigned from a random range.