Introduce BitcoinFixture
BitcoinFixture is a shared fixture for all integration tests that uses
reference counting to avoid having the private key of the mining account
in more than one bitcoinj at the same time.  Apart from that, the fixture
is used to fund the wallets of the peer to participate on the integration
tests.

To speed the tests, the bitcoin fixture is shared among all of them. To
do so, it is initialized as soon as BitcoinFixture.instance lazy val is
used and freed via JVM shutdown hook.