Fix a race condition in `OrderActor`
There is a race condition in `OrderActor`. Once the order is offline, the
actor proceeds indicating the order submission actor to keep submitting the
order to the market. Then, the following sequence of actions may happen.

* The order submission actor sends the order to the broker. The broker receives
it and sends its confirmation and an order match message.
* The order submission actor receives the confirmation from the broker and
prepares to send the `InMarket` message to the `OrderActor`, but before it gets
to do so...
* The order match message arrives at the `OrderActor`. Since the order is still
offline, it is not ready to process the message and ignores it. The order
match is lost.
* The order submmission actor finally sends `InMarket` to the `OrderActor`. Now
it transits to the `waitingForMatch` state, but the order match has already
been sent and ignored. The exchange never begins.

This commit fixes this problem by considering that `offline` and
`waitingForMatch` as the same state.