Replace `Connect` message by `Join` in `MessageGateway`
We want the message gateway to join the network rather than connect to it.
The difference between these two terms is that connect implies a network channel
to be established while join means that it will attempt to form part of the
network, establishing as many channels as required and reconnecting them as
they are lost.

This change replaces the term _connect_ by _join_ in message gateway classes.