Enrich `AskPattern` with functions to expect success or error
There is a common pattern when asking an actor: expect either a success or an
failure response message. In some parts of the code, we are dealing with that
by defining a marking trait to instantiate template function `withReply()` and
then using a collect/map function. In some other parts, we are just invoking
`withReply[Any]`.

This commit adds two more functions to `AskPattern` object, `withReplyOrError()`
and `withImmediateReplyOrError()`. Both functions require two generic
parameters: the expected success and the expected failure message types. In
addition, it's possible to pass a function that converts the failure message
into a `Throwable` to set the future failure (by default throws a
`RuntimeException` indicating that a failure message was received).