Introduce `NodeId` and `BrokerId`
This commit introduces a new sealed trait `NodeId` used to identify nodes in
the Coinffeine network. It has two implementations: our old-friend `PeerId` and
a new one, `BrokerId`, used to identify univoquely the broker.

All the code has been refactored to use `NodeId` where the ID of any node is
expected instead of just the ID of a regular peer.