Move initialization of `MessageForwarder` to its constructor
`MessageForwarder` instances are designed to forward a single message in all
its lifecycle. That means they could be simplified by moving its initialization
from the `Forward` message receival to its constructor. That's what this commit
does.

After this change, the `changeSubscriptions` flag of `relayMessage()` in
`GatewayProbe` are no longer necessary and removed.