Make `BuyerSellerCoordinationTest` use `LinkedMessageGateways`
The mocked communication used by this tests if both actors are sending and
receiving their messages. But in case they are using a message forwarder the
test would fail because the responses are directly sent to the main actor and
not its forwarder.

This change adapts the test to use the new `LinkedMessageGateways` and therefore
be ready to support message forwarders in both actors.