Use message forwarding in `SellerMicropaymentChannel`
Instead of forwarding the signatures and setting its timeouts manually, after
this change `SellerMicropaymentChannel` uses `MessageForwarder` utility to do
it. Since all signature sent to the buyer requires confirmation, the buyer
is changed as well to respond with a `MicropaymentChannelClosed` message when
the last signature is received.

This commit removes the only test in `SellerMicroPaymentChannelActorFailureTest`
since it was meningless after this change. The class is not removed since more
tests will be added in the near future (removing the class to create it
again tomorrow is not very wise).