Change how crosses are cleared on the order book
Instead of automatically clearing order crosses, make them explicitly
start a handshake that can be completed successfully or be cancelled.
Make also orders already on handshake not being matched against other
orders (sequential model of matching) but preserve their priority.

In the case of a failed handshake, the orders of the misbehaving peers
will be cancelled but the ones of honest peers will keep their priority.