Introduce settings as replacement of config
Consuming `Config` objects to extract the settings is not safe. The keys are
spread all over the code, the values have to be converted into the expected
type, they have to be merged to obtain an aggregated config view, ...

This change introduces the concept of settings. The `SettingsProvider` now
provides settings instead of config. The settings provide specific configuration
for each aspect of the app in the form of a case class (no more type unsafe
key-value maps). The canonical implementation of `SettingsProvider` is aimed
to read Typesafe `Config` objects and convert them into settings.