Adjust constants to have successful tests
As now the step size depends on the amounts exchanged, ExchangeTest
was producing a 500-steps exchange. That is an issue since our lock time
is measured on blocks and that means minutes not hours with our current
testnet setup.

To have a reasonable result, this patch increments the lock time to 18
and reduces the exchanged amount to 100 EUR.