Fix CrossedOrdersTest
Use the same price for both orders as there is an issue with reserving
funds when the price is better than the limit for the buyer and
increase the amounts the peers have to avoid stalling the orders.

Also, fix the time available for performing the exchange handshake and
improve logging about blocked amounts.