Allow dropping reference config items while saving
A new boolean flag in `saveUserConfig()` can be used to indicate whether the
config items that match those of reference config have to be dropped away. A
config item is considered to match a reference one if both its key and value
matches. In other words, if the user references a new value for a known config
item, it shall be part of user configuration.