Simplify `PeerGroupComponent`
The current interface of `PeerGroupComponent` considers that the implementers
must use an `AbstractBlockChain` to create the peer group. That's an
implementation detail that should not be present in the trait.

This change replaces the interface of `PeerGroupComponent` by a one that
just retrieves the peer group however it's created. That leads to move
its implementation from `IntegrationTestNetworkComponent` to a new trait
`SimplePeerGroupComponent`.