Add a new data model based on reactive properties
This new data model is designed to be consumed by both the actor system and
the Coinffeine API. The former would have the ability to mutate the properties
while the latter is only able to read them.

This data model allows the direct access to the Coinffeine data from the API
without interacting with the actors. That reduces the response latencies,
reduces the data fetching complexity and removes the risk of unresponsive actors
unable to satisfy the data retrieval request.