Fix a bug that caused multiple order submitters to coexist
In some circumstances, the `MarketSubmissionActor` was spawning a new
submitter that coexisted with the last one before its death. This causes an
actor naming clashing and could derive in more serious problems when old and
new submitters fight to deliver its orders.

This commit fixes the problem by waiting for the previous submitter (if any)
to terminate before spawning a new one.