Prevent issues with old messages in `OkPayProcessorActorTest`
We are using the same test probe to send and receive the responses of the
different test cases. If some test case finishes but the actor under test is
not terminated, it could send messages to the test probe that would be received
while having different expectations in another test case.

This commit fixes this heisenbug (CF-86) by using a test probe in the fixture,
which ensures that the requester is different from one test case to another.