Provide a popup widget for bitcoin balance widget
To do that, we had moved `WalletBalanceWidget` to an specific package and it
has been splitted into two child classes: one for Bitcoin and another one for
fiat. Their parent class `WalletBalanceWidget` now must be mixed with a
`PopUpContent` trait that provides the contents for the popup it manages. If
none is provided (the case of fiat widget), the popup is never shown.