Fix a bug that prevented to load bitcoin settings from scratch
Bitcoin settings depends on `coinffeine.bitcoin.walletFile` to be defined in
the configuration. This property, which is platform dependent, was not defined
in the static reference configuration for obvious reasons. It was aimed to set
the user configuration by establishing the wallet file upon app setup. But
that's not possible since the settings must be loaded before modified, and
the config cannot be loaded due to the wallet file key is missing.

This change fixes the issue by considering that the reference config comes
from the static files and additional config items that are calculated using
code.