Reload OKPay client configuration
Use options on the corresponding SettingsMapping to represent 
unconfigured payment processors and, therefore,  use Option[AccountId] 
in several other interfaces. Also, remove default values from the 
application.conf at coinffeine-gui.

Extract OkPayWebService class from OkPayWebServiceClient. The 
heavyweight initialization is done by the former and we can have
a fresh instance of the latter with the last version of the settings.