Fix missing commitment ACK retransmission from peer
It seems that `HandshakeActor` is not considering the chance of
`CommitmentNotificationAck` to be lost. As I read in the last build log, the
chaos message gateway is dropping that message sent from the peer to the broker.
Then, the broker insists by sending `CommitmentNotification` message again and
again, which is never received by the peer. The broker never considers the
handshake is done, and the test fails. Fixed by resending
`CommitmentNotificationAck` from the peer.