Fix not shown wizard
Since we are handling the config before wizard is launched, the criteria used to
determine whether the wizard must be shown is not valid anymore. This change
replaces that criteria, which just checks whether the user config exists, by
a check of the properties handled by the wizard are set.