Fix a bug in `MarketInfoActor` that causes a fail to retrieve market stats
In `MarketInfoActor`, if some of the pending requests fails twice due to
message gateway is unable to send the message, the request remains in the pool
of pending requests not retrying anymore and preventing any other request to
replace it.

This change configures the request forwarder to retry forever until success.
That fixes the bug.