Fix lack of informatio in order submission form
The fiat currency amount requested in the order submission form is actually
the price per bitcoin. The UI is not clear on that. It would be great to have
another box or label that indicates the total amount of fiat that would be
exchanged.

This change adds additional labels that react against amounts amounts text
fields printing what's the total amount of fiat for the new order.