Consider only tx output #0 when computing spend deposit candidates
Deposit transactions use to have two outputs. One for the exchange, and another
to the change address. Since we were not considering this, any transaction
spending the change output #1 is considered as `DepositUnlock` instead of
`OutFlow`.

This commit fixes that by filtering the outpoints with index #0 when calculating
the spend candidates.